Use blob download instead of data-uri to enable large downloads
This was already being used for sb3s and sprite3s, but the code was repeated in different places. Make a single helper that downloads blobs, and make all download paths use it.
