--- src/lib/download-url.js
@@ -1,13 +1,0 @@
-export default (filename, url) => {
-    const pom = document.createElement('a');
-    pom.setAttribute('href', url);
-    pom.setAttribute('download', filename);
-
-    if (document.createEvent) {
-        const event = document.createEvent('MouseEvents');
-        event.initEvent('click', true, true);
-        pom.dispatchEvent(event);
-    } else {
-        pom.click();
-    }
-};
